feat(playground-ui): add on-the-fly sub-agent creation via SideDialog
Extract agent creation form into reusable AgentCreateContent component
in playground-ui, then add a "Create" button + SideDialog to the
Sub-Agents section so users can create sub-agents without leaving the
agent editor. Uses hideSubAgentCreate prop to prevent recursive nesting.
