+++ .changeset/shaggy-badgers-move.md
@@ -1,0 +1,5 @@
+---
+'@mastra/playground-ui': minor
+---
+
+Added ability to create sub-agents on-the-fly via a SideDialog in the Sub-Agents section of the agent editor
