+++ packages/playground-ui/src/domains/agents/components/agent-edit-page/agent-edit-sidebar.tsx
@@ -79,6 +79,7 @@ interface AgentEditSidebarProps {
   formRef?: RefObject<HTMLFormElement | null>;
   mode?: 'create' | 'edit';
   readOnly?: boolean;
+  hideSubAgentCreate?: boolean;
 }
 
 export function AgentEditSidebar({
@@ -89,6 +90,7 @@ export function AgentEditSidebar({
   formRef,
   mode = 'create',
   readOnly = false,
+  hideSubAgentCreate,
 }: AgentEditSidebarProps) {
   const {
     register,
@@ -234,6 +236,7 @@ export function AgentEditSidebar({
                 error={errors.agents?.root?.message}
                 currentAgentId={currentAgentId}
                 readOnly={readOnly}
+                hideCreateButton={hideSubAgentCreate}
               />
               <ScorersSection control={control} readOnly={readOnly} />
               <MemorySection control={control} setValue={form.setValue} readOnly={readOnly} />
