+++ packages/playground-ui/src/domains/agents/index.tsx
@@ -24,3 +24,4 @@ export * from './components/agent-layout';
 export * from './components/agent-cms-blocks';
 export * from './components/agent-edit-page/utils/form-validation';
 export * from './components/agent-edit-page';
+export * from './components/agent-create-content';
